"Added a new Python module in naive_agi/framework"
A new Python module has been added to the 'framework' package under the 'naive_agi' project. Currently, the module is empty as indicated by the index hash 'e69de29', which corresponds to an empty file in git. Future updates to this module will contain specific functionalities. Please note this change as it's an initialization of a new component in our project.